+++ pyslvs/__init__.py
@@ -15,9 +15,6 @@ from .tinycadlib import (
     pxy, ppp, plap, pllp, plpp, palp, vpoint_dof, expr_solving, uniform_path,
     uniform_four_bar,
 )
-from .planar_linkage import (
-    FMatch, norm_path, curvature, derivative, path_signature, cross_correlation,
-)
 from .expression_parser import (
     color_names, color_rgb, parse_params, parse_pos, parse_vpoints,
     parse_vlinks, edges_view, graph2vpoints, PointArgs, LinkArgs,